% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chebyshev.R
\name{chebyshev}
\alias{chebyshev}
\title{Chebyshev distance matrice between points}
\usage{
chebyshev(x, y, output = "matrix")
}
\arguments{
\item{x}{Numeric vector, or a matrix with size n x m.}

\item{y}{(optional)Numeric vector; must be the same length as x. If no input of y,
please specify all other parameters to avoid confusion.}

\item{output}{Format of output, either matrix or dist object. Default is matrix.}
}
\value{
Chebyshev distance matrix or object, max(abs(x - y)) between points
}
\description{
Computes the Chebyshev distance between n points in m dimensions
,or two points with same dimensions.
}
\examples{
chebyshev(x = matrix(c(1, 2, 3, 4), 2, 2), output = "dist")

chebyshev(rnorm(10), rnorm(10))

}
