% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_dis.R
\name{corr_dis}
\alias{corr_dis}
\title{Correlation distance matrice between points}
\usage{
corr_dis(x, y, output = "matrix")
}
\arguments{
\item{x}{Numeric vector, or a matrix with size n x m.}

\item{y}{(optional)Numeric vector; must be the same length as x. If no input of y,
please specify all other parameters to avoid confusion.}

\item{output}{Format of output, either matrix or dist object. Default is matrix.}
}
\value{
Correlation distance matrix or object, centered form of cosine distance between points
}
\description{
Computes the correlation distance between n points in m dimensions
,or two points with same dimensions.
}
\examples{
corr_dis(x = matrix(c(1, 2, 3, 4), 2, 2), output = "dist")

corr_dis(rnorm(10), rnorm(10))

}
