% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_dis.R
\name{cos_dis}
\alias{cos_dis}
\title{Cosine distance matrice between points}
\usage{
cos_dis(x, y, output = "matrix")
}
\arguments{
\item{x}{Numeric vector, or a matrix with size n x m.}

\item{y}{(optional)Numeric vector; must be the same length as x. If no input of y,
please specify all other parameters to avoid confusion.}

\item{output}{Format of output, either matrix or dist object. Default is matrix.}
}
\value{
Cosine distance matrix or object calculated by 1 - dot(x, y) / (norm(x) * norm(y)) between points
}
\description{
Computes the Cosine distance between n points in m dimensions
,or two points with same dimensions.
}
\examples{
cos_dis(x = matrix(c(1, 2, 3, 4), 2, 2), output = "dist")

cos_dis(rnorm(10), rnorm(10))

}
