% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean distance matrice between points}
\usage{
euclidean(x, y, w = 1, output = "matrix")
}
\arguments{
\item{x}{Numeric vector, or a matrix with size n x m.}

\item{y}{(optional)Numeric vector; must be the same length as x. If no input of y,
please specify all other parameters to avoid confusion.}

\item{w}{Numeric value greater than 0, account for weight. Default is 1.}

\item{output}{Format of output, either matrix or dist object. Default is matrix.}
}
\value{
Euclidean distance matrix or object calculated by sqrt(sum((x - y) ^ 2)) between points
}
\description{
Computes the Euclidean distance between n points in m dimensions
,or two points with same dimensions.
}
\examples{
euclidean(x = matrix(c(1, 2, 3, 4), 2, 2), output = "dist")

euclidean(rnorm(10), rnorm(10), w = 2)

}
